requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
python-dotenv==1.0.1
pymongo==4.6.2
Werkzeug==3.0.1
WTForms==3.1.2
bcrypt==4.1.2
email-validator==2.1.0.post1
dnspython==2.7.0
Jinja2==3.1.3
MarkupSafe==2.1.5
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
beautifulsoup4==4.12.3
nltk==3.8.1
requests==2.31.0
certifi==2025.1.31
openai==1.12.0 